requirements.txt
aiohttp==3.8.3
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
discord.py==2.0.1
discord==2.0.0
multidict==6.0.2
oauthlib==3.2.2
osrparse==6.0.1
ossapi==2.4.2
pip==22.3
python-dotenv==0.21.0
requests-oauthlib==1.3.1
requests==2.28.1
setuptools==65.5.1
typing-extensions==4.4.0
typing-utils==0.1.0
urllib3==1.26.12
yarl==1.8.1
